Add files via upload
I tried make it but I don't know how to designate closestPlant to right plot, plant, and collect that are tagged. First I divided 'FindCloeset~' so that it can proceed sequentially. And I added 'BuyPlotsAI' for buying low cost land first. And I also divided 'MeshRaycast~2' so that it can detect "Buy"(tag) emptyplot that need to buy and "Plant","Collect"(tag) plot that player already bought. And I made 'orderFindCloest'  in the form of buying land first, planting seeds on empty land, and collecting money when money is scarce. What we need to is to set closestPlant to plot, plant, and collect. And I don't know it will work because I didn't saw it worked because AI choose plot randomly because closestPlant is not designated